feat: token type pill on send funds token picker
